Cleanup the utils module
