+++ src/utils/buf.rs
@@ -1,0 +1,7 @@
+//! Helpers for dealing with byte buffers.
+
+/// Alias for [u8].
+pub type Byte = u8;
+
+/// Alias for byte slices.
+pub type Buf<'b> = &'b [Byte];
